Refactor: Move ClearlyDefinedResponse to a separate file
Relocated the ClearlyDefinedResponse class from ClearlyDefinedClient.cs to a new ClearlyDefinedResponse.cs file. This improves code organization and separation of concerns, making the project structure cleaner and easier to maintain.
